Version Bump
I changing the way I name version.

For blender 3.6 version will be 3.6.x
For blender 4.0 version will be 4.0.x

For example if belnder 4.1 comes out cats will now be 4.1.x once a 4.1 version is released.

This is to future proof for when blender decides to break everything again in a new major version.