Possible fix to #44
This is an possible fix for #44 this adds a new checkmark to optimization > bones called Experimental: Keep bones with _twist this allows the user to add _twist or Twist to bones and the button  Remove Zero Weight bones will not remove these bones if this is checked.

This is marked experimental for now as I not sure if this will cause other issues and if there is bugs with this.