Shapekey Translation Export to CSV now uses file name.
When exporting translations to csv for shapekeys it will save as the current blend file name, if there isn't a save file it will error out asking the user to save the blend file. If the option is disabled translations will work regardless of the file being saved or not.

Part of request #51